+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/failure/Extensions.kt
@@ -19,6 +19,7 @@ package im.vector.matrix.android.api.failure
 import im.vector.matrix.android.api.extensions.tryThis
 import im.vector.matrix.android.internal.auth.registration.RegistrationFlowResponse
 import im.vector.matrix.android.internal.di.MoshiProvider
+import java.io.IOException
 import javax.net.ssl.HttpsURLConnection
 
 fun Throwable.is401() =
@@ -32,6 +33,7 @@ fun Throwable.isTokenError() =
 
 fun Throwable.shouldBeRetried(): Boolean {
     return this is Failure.NetworkConnection
+            || this is IOException
             || (this is Failure.ServerError && error.code == MatrixError.M_LIMIT_EXCEEDED)
 }